% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpp.R
\name{as.data.frame.mpp_sim}
\alias{as.data.frame.mpp_sim}
\title{as.data.frame implementation for mmp_sim}
\usage{
\method{as.data.frame}{mpp_sim}(x, ..., discard_initial = FALSE)
}
\arguments{
\item{x}{Simulation of Marked Point Process by rmpp.}

\item{...}{not used.}

\item{discard_initial}{Should inital state and start time be discarded.}
}
\value{
data.frame
}
\description{
as.data.frame implementation for mmp_sim
}
