% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity_builders.R
\name{build_probs}
\alias{build_probs}
\title{build_probs}
\usage{
build_probs(x, type = "mpp")
}
\arguments{
\item{x}{matrix of intensities.}

\item{type}{If you intend to use rates for rmarkov or rsemimarkov you must
specify which one. Either 'markov' or 'semimarkov'.}
}
\value{
list of functions
}
\description{
Build transitions probabilities from matrix of intensities to be used with
rmpp.
}
