% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity_builders.R
\name{build_rates}
\alias{build_rates}
\title{build_rates}
\usage{
build_rates(x)
}
\arguments{
\item{x}{matrix}
}
\value{
list of functions
}
\description{
Constructs rates from a matrix of intensities to be used with rmpp.
}
