% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity_builders.R
\name{complete_intensity_matrix}
\alias{complete_intensity_matrix}
\title{complete_intensity_matrix}
\usage{
complete_intensity_matrix(x, type = "mpp")
}
\arguments{
\item{x}{intensity matrix. Formally a list of functions.}

\item{type}{Can be either markov, semimarkov or mpp.}
}
\value{
list
}
\description{
Completes implicit NULL's in intensity matrix.
}
