% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity_builders.R
\name{dominate}
\alias{dominate}
\title{dominate}
\usage{
dominate(fns)
}
\arguments{
\item{fns}{A list of functions}
}
\value{
A list of functions
}
\description{
Automatic domination of intensities used in rmpp.
}
