% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpp.R
\name{print.mpp_sim}
\alias{print.mpp_sim}
\title{print method for rmpp simulations.}
\usage{
\method{print}{mpp_sim}(x, ..., n = 1, digits = 2)
}
\arguments{
\item{x}{Simulation of Marked Point Process by rmpp.}

\item{...}{not used.}

\item{n}{Number of paths to be displayed}

\item{digits}{Amount of digits to be displayed}
}
\value{
Text display
}
\description{
print method for rmpp simulations.
}
