% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpp.R
\name{summary.mpp_sim}
\alias{summary.mpp_sim}
\title{summary method for rmpp simulations.}
\usage{
\method{summary}{mpp_sim}(x, discard_inital = FALSE)
}
\arguments{
\item{x}{Simulation of Marked Point Process by rmpp.}

\item{discard_intial}{Should inital state and start time be discarded.}
}
\value{
data.frame
}
\description{
summary method for rmpp simulations.
}
