% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latrix.R
\name{latrix}
\alias{latrix}
\title{latrix}
\usage{
latrix(data = NA, nrow = 1, ncol = 1, byrow = FALSE, dimnames = NULL)
}
\arguments{
\item{data}{data}

\item{nrow}{number of rows}

\item{ncol}{number of columns}

\item{byrow}{should data be arranged by row instead of by column?}

\item{dimnames}{A dimnames attribute for the matrix.}
}
\value{
matrix
}
\description{
A wrapper around matrix in order to create a list matrix.
}
\examples{
A 2 x 2 empty list mat
latrix(nrow = 2, ncol = 2)
}
