% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpp.R
\name{rmpp}
\alias{rmpp}
\title{rmpp Simulate paths of marked point process on a finite interval.}
\usage{
rmpp(
  n,
  rates,
  probs,
  t0,
  tn,
  y0 = 1L,
  mark_end = NULL,
  drates = NULL,
  limit = 1e+08,
  ...
)
}
\arguments{
\item{n}{Number of paths to be simulated}

\item{rates}{List of functions returning outflow of rates from each state.}

\item{probs}{List of functions that return probability of jumping to each
state.}

\item{t0}{Starting time. Must have length 1 or same length as number of
paths.}

\item{tn}{Stoping time. Must have length 1 or same length as number of paths.}

\item{y0}{Starting marks. Must have length 1 or same length as number of
paths.}

\item{mark_end}{If end time tn is reached before absorbing state then it is
marked by a mark specified by mark_end.}

\item{drates}{List of functions that dominates each outflow rate in 'rates'.
Intensities in 'rates' arguments are dominated by default using the
optimize routine of the stats package. This is noticeably slower than
dominating them manually and it is recommended to do so as optimize will not
always obtain an exact upper bound leading to warnings.}

\item{limit}{Jump limit. If this limit is reached the simulation of the
particular path will stop. This is a safety measure.}

\item{...}{Additional arguments to be passed to the rate functions. This
could be baseline parameters such as age, gender etc. They must be
numerics, doubles or integers.}
}
\value{
list of simulated paths
}
\description{
rmpp Simulate paths of marked point process on a finite interval.
}
